requirements.txt
flask
pyyaml
pyperclip
gunicorn
eventlet